Optimization, for real this time
Removed the new global that was tracking existence of entities, finally realized they do that themselves.

With this, no new event handlers are needed, just a cache of the entities that are going into chunks/jobs.

Once a job is active, each time it cycles through while the job is in the starting state, check the job's cached entities to make sure they're still valid. If any are not valid, remove them and adjust the job's required items. Finally, remove the items_requested sub_state, to force the worker to re-do its logistic requests.